[#1942] Factor out session tracking logic in testing backend
